chore(api): move string-comparison-service.js to evaluation scope
