+++ api/src/evaluation/domain/services/solution/solution-service-qroc.js
@@ -1,6 +1,5 @@
 import lodash from 'lodash';
 
-import { validateAnswer } from '../../../../../lib/domain/services/string-comparison-service.js';
 import { applyPreTreatments, applyTreatments } from '../../../../../lib/domain/services/validation-treatments.js';
 import { AnswerStatus } from '../../../../shared/domain/models/AnswerStatus.js';
 import {
@@ -9,6 +8,7 @@ import {
   splitIntoWordsAndRemoveBackspaces,
 } from '../../../../shared/infrastructure/utils/string-utils.js';
 import { getEnabledTreatments, useLevenshteinRatio } from '../services-utils.js';
+import { validateAnswer } from '../string-comparison-service.js';
 const CHALLENGE_NUMBER_FORMAT = 'nombre';
 const { every, isEmpty, isString, map } = lodash;
