+++ api/src/evaluation/domain/services/string-comparison-service.js
@@ -1,7 +1,7 @@
 import levenshtein from 'fast-levenshtein';
 import _ from 'lodash';
 
-import { LEVENSHTEIN_DISTANCE_MAX_RATE } from '../../../src/shared/domain/constants.js';
+import { LEVENSHTEIN_DISTANCE_MAX_RATE } from '../../../shared/domain/constants.js';
 
 function isOneStringCloseEnoughFromMultipleStrings(inputString, references, MAX_ACCEPTABLE_RATIO) {
   return getSmallestLevenshteinRatio(inputString, references) <= MAX_ACCEPTABLE_RATIO;
