+++ api/tests/evaluation/unit/domain/services/string-comparison-service_test.js
@@ -2,12 +2,12 @@ import {
   getSmallestLevenshteinDistance,
   getSmallestLevenshteinRatio,
   isOneStringCloseEnoughFromMultipleStrings,
-} from '../../../../lib/domain/services/string-comparison-service.js';
+} from '../../../../../src/evaluation/domain/services/string-comparison-service.js';
 import {
   areTwoStringsCloseEnough,
   getLevenshteinRatio,
-} from '../../../../src/shared/domain/services/string-comparison-service.js';
-import { expect } from '../../../test-helper.js';
+} from '../../../../../src/shared/domain/services/string-comparison-service.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Service | Validation Comparison', function () {
   describe('getSmallestLevenshteinDistance', function () {
